The Guessing game C++
A quick test of C++, using the language to create a basic DOS like number guessing game. As I'm interested in the unreal engine (which runs on C++) I thought it would be beneficial to gain some basic knowledge of the language before learning blue print. On top of this the basics could give me a stepping stone to other programming languages like C# (which is used in Unity)

I created two 'integers', one for the hidden target number (CorrectNum) and the other for input from the player to create a means to guess and compare values (GuessNum).

Then I created a 'while' loop with various 'if' statements. I made five outcomes, one for guessing too high but between 0 and 100, another for being too low but between 0 and 100, and two for being completely outside the 0 to 100 range. If the number was guessed incorrectly the player would be told if they were too high or too low, and the 'while' loop would repeat until the number was guessed, where the player would be congratulated and the application would shutdown (using the 'bool' 'Correct').

I originally learnt to do this from a tutorial some time ago, but reversed engineered my previous attempt to test other methods of writing the code, and to truly understand each part.